Added At Risk and Pregnancy statistics to FRED through the use of Age_Maps.
Pregnancy is added to Population, and then affects the agent's Demographics.
At Risk is added to Strain, as it is a strain dependent thing, and then affects the agent's Health.

There are also some changes in here to free up memory allocated by Spread, Strain, Person, Timestep_Map and Population.
I also added default constructors that put in values that will fail quality assurance if they are not then initalized to
many of the classes, and added default destructors for many.

Did some work as well to make many of the classes be more conformant to the Google/FRED standard :) (for alona)

-- Shawn Brown < stbrown@psc.edu>


git-svn-id: https://redmine.hindsight.psc.edu/svn/fred/trunk/FRED@188 b93b844c-2574-459d-9dfb-c4b8cb2af1a1